Separating the ShowSession, Reservation and Ticket models into a separate 'reservations' application
