+++ planetarium/admin.py
@@ -1,14 +1,6 @@
 from django.contrib import admin
-from planetarium.models import (
-    ShowTheme,
-    AstronomyShow,
-    Reservation,
-    ShowSession,
-    Ticket,
-)
+from planetarium.models import ShowTheme, AstronomyShow
 
 admin.site.register(ShowTheme)
 admin.site.register(AstronomyShow)
-admin.site.register(Reservation)
-admin.site.register(ShowSession)
-admin.site.register(Ticket)
+
