+++ planetarium/models.py
@@ -1,10 +1,7 @@
 import os
 import uuid
-from django.conf import settings
-from django.core.exceptions import ValidationError
 from django.db import models
 from django.utils.text import slugify
-from domes.models import PlanetariumDome
 
 
 class ShowTheme(models.Model):
@@ -38,93 +35,3 @@ class AstronomyShow(models.Model):
 
     class Meta:
         ordering = ["id"]
-
-
-class Reservation(models.Model):
-    created_at = models.DateTimeField(auto_now_add=True)
-    user = models.ForeignKey(
-        settings.AUTH_USER_MODEL, on_delete=models.CASCADE
-    )
-
-    def __str__(self):
-        return f"{str(self.created_at)} by: {self.user}"
-
-    class Meta:
-        ordering = ["-created_at"]
-
-
-class ShowSession(models.Model):
-    astronomy_show = models.ForeignKey(
-        AstronomyShow,
-        on_delete=models.CASCADE
-    )
-    planetarium_dome = models.ForeignKey(
-        PlanetariumDome,
-        on_delete=models.CASCADE
-    )
-    show_time = models.DateTimeField()
-
-    class Meta:
-        ordering = ["show_time"]
-
-    def __str__(self):
-        return self.astronomy_show.title + " " + str(self.show_time)
-
-
-class Ticket(models.Model):
-    show_session = models.ForeignKey(
-        ShowSession, on_delete=models.CASCADE, related_name="tickets"
-    )
-    reservation = models.ForeignKey(
-        Reservation, on_delete=models.CASCADE, related_name="tickets"
-    )
-    row = models.IntegerField()
-    seat = models.IntegerField()
-
-    @staticmethod
-    def validate_ticket(row, seat, planetarium_dome, error_to_raise):
-        for (ticket_attr_value,
-             ticket_attr_name,
-             planetarium_dome_attr_name) in [
-            (row, "row", "rows"),
-            (seat, "seat", "seats_in_row"),
-        ]:
-            count_attrs = getattr(planetarium_dome, planetarium_dome_attr_name)
-            if not (1 <= ticket_attr_value <= count_attrs):
-                raise error_to_raise(
-                    {
-                        ticket_attr_name: f"{ticket_attr_name} "
-                        f"number must be in available range: "
-                        f"(1, {planetarium_dome_attr_name}): "
-                        f"(1, {count_attrs})"
-                    }
-                )
-
-    def clean(self):
-        Ticket.validate_ticket(
-            self.row,
-            self.seat,
-            self.show_session.planetarium_dome,
-            ValidationError,
-        )
-
-    def save(
-        self,
-        force_insert=False,
-        force_update=False,
-        using=None,
-        update_fields=None,
-    ):
-        self.full_clean()
-        return super(Ticket, self).save(
-            force_insert, force_update, using, update_fields
-        )
-
-    def __str__(self):
-        return (
-            f"{str(self.show_session)} (row: {self.row}, seat: {self.seat})"
-        )
-
-    class Meta:
-        unique_together = ("show_session", "row", "seat")
-        ordering = ["row", "seat"]
