+++ planetarium/urls.py
@@ -3,15 +3,11 @@ from rest_framework import routers
 from planetarium.views import (
     ShowThemeViewSet,
     AstronomyShowViewSet,
-    ShowSessionViewSet,
-    ReservationViewSet,
 )
 
 router = routers.DefaultRouter()
 router.register("show-themes", ShowThemeViewSet)
 router.register("astronomy-shows", AstronomyShowViewSet)
-router.register("show-sessions", ShowSessionViewSet)
-router.register("reservations", ReservationViewSet)
 
 urlpatterns = [
     path("", include(router.urls))
