+++ planetarium/views.py
@@ -1,33 +1,14 @@
-from datetime import datetime
-from django.db.models import Count, F
 from drf_spectacular.types import OpenApiTypes
 from drf_spectacular.utils import extend_schema, OpenApiParameter
-from rest_framework import mixins, viewsets, status
+from rest_framework import mixins, status
 from rest_framework.decorators import action
-from rest_framework.pagination import PageNumberPagination
-from rest_framework.permissions import IsAuthenticated, IsAdminUser
+from rest_framework.permissions import IsAdminUser
 from rest_framework.response import Response
 from rest_framework.viewsets import GenericViewSet
-
-from planetarium.models import (
-    ShowTheme,
-    AstronomyShow,
-    ShowSession,
-    Reservation,
-)
+from planetarium.models import ShowTheme, AstronomyShow
 from planetarium.permissions import IsAdminOrIfAuthenticatedReadOnly
-from planetarium.serializers import (
-    ShowThemeSerializer,
-    AstronomyShowSerializer,
-    AstronomyShowListSerializer,
-    AstronomyShowDetailSerializer,
-    AstronomyShowImageSerializer,
-    ShowSessionListSerializer,
-    ShowSessionSerializer,
-    ShowSessionDetailSerializer,
-    ReservationSerializer,
-    ReservationListSerializer,
-)
+from planetarium.serializers import ShowThemeSerializer, AstronomyShowSerializer, AstronomyShowListSerializer, \
+    AstronomyShowDetailSerializer, AstronomyShowImageSerializer
 
 
 class ShowThemeViewSet(
@@ -123,101 +104,3 @@ class AstronomyShowViewSet(
     )
     def list(self, request, *args, **kwargs):
         return super().list(request, *args, **kwargs)
-
-
-class ShowSessionViewSet(viewsets.ModelViewSet):
-    queryset = (
-        ShowSession.objects.select_related(
-            "astronomy_show",
-            "planetarium_dome"
-        ).order_by("show_time")
-        .annotate(
-            tickets_available=(
-                F("planetarium_dome__rows")
-                * F("planetarium_dome__seats_in_row")
-                - Count("tickets")
-            )
-        )
-    )
-    serializer_class = ShowSessionListSerializer
-    permission_classes = (IsAdminOrIfAuthenticatedReadOnly,)
-
-    def get_queryset(self):
-        date = self.request.query_params.get("date")
-        astronomy_show_id_str = self.request.query_params.get("astronomy_show")
-
-        queryset = self.queryset
-
-        if date:
-            date = datetime.strptime(date, "%Y-%m-%d").date()
-            queryset = queryset.filter(show_time__date=date)
-
-        if astronomy_show_id_str:
-            queryset = queryset.filter(
-                astronomy_show_id=int(astronomy_show_id_str)
-            )
-
-        return queryset
-
-    def get_serializer_class(self):
-        if self.action == "list":
-            return ShowSessionListSerializer
-
-        if self.action == "retrieve":
-            return ShowSessionDetailSerializer
-
-        return ShowSessionSerializer
-
-    @extend_schema(
-        parameters=[
-            OpenApiParameter(
-                "astronomy_show",
-                type=OpenApiTypes.INT,
-                description=(
-                    "Filter by astronomy show id "
-                    "(ex. ?astronomy_show=2)"
-                ),
-            ),
-            OpenApiParameter(
-                "date",
-                type=OpenApiTypes.DATE,
-                description=(
-                    "Filter by datetime of show time "
-                    "(ex. ?date=2022-10-23)"
-                ),
-            ),
-        ]
-    )
-    def list(self, request, *args, **kwargs):
-        return super().list(request, *args, **kwargs)
-
-
-class ReservationPagination(PageNumberPagination):
-    page_size = 10
-    max_page_size = 100
-
-
-class ReservationViewSet(
-    mixins.ListModelMixin,
-    mixins.CreateModelMixin,
-    GenericViewSet,
-):
-    queryset = Reservation.objects.prefetch_related(
-        "tickets__show_session__astronomy_show",
-        "tickets__show_session__planetarium_dome",
-    )
-    serializer_class = ReservationSerializer
-    pagination_class = ReservationPagination
-    permission_classes = (IsAuthenticated,)
-
-    def get_queryset(self):
-        return self.queryset.filter(user=self.request.user)
-
-    def get_serializer_class(self):
-        if self.action == "list":
-            return ReservationListSerializer
-
-        return ReservationSerializer
-
-    def perform_create(self, serializer):
-        serializer.save(user=self.request.user)
