+++ planetarium_service/settings.py
@@ -47,6 +47,7 @@ INSTALLED_APPS = [
     "drf_spectacular",
     "debug_toolbar",
     "domes",
+    "reservations",
     "planetarium",
     "user",
 ]
